requirements.txt
pandas
numpy
scikit-learn
gensim
spacy
nltk
streamlit
python-dotenv
joblib
scipy